+++ libs/platform/src/lib/table-helpers/services/table.service.ts
@@ -354,7 +354,8 @@ export class TableService {
             page: {
                 pageSize: 0,
                 currentPage: 1
-            }
+            },
+            scrollTopPosition: 0
         };
     }